Laravel Breeze Auth

+++ meu-app-laravel/routes/web.php
@@ -8,32 +8,25 @@ use App\Http\Controllers\{
     ViaCepController
 };
 
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
-
-Route::get("/user/team/{id}", [TeamController::class, 'show'])->name('team.show');
-
-Route::get('/posts', [PostController::class, 'index'])->name('posts.index');
-Route::get('/users/{id}/posts', [PostController::class, 'show'])->name('posts.show');
-
-Route::delete("/users/{id}/delete", [UserController::class, "destroy"])->name("users.destroy");
-Route::put("/users/{id}", [UserController::class, "update"])->name("users.update");
-Route::get("/users/{id}/edit", [UserController::class, "edit"])->name("users.edit");
-Route::post("/users", [UserController::class, "store"])->name("users.store");
-Route::get("/", [UserController::class, "index"])->name("users.index");
-Route::get("/users/create", [UserController::class, "create"])->name("users.create");
-Route::get("/users/{id}", [UserController::class, "show"])->name("users.show");
+require __DIR__.'/auth.php';
 
 
-//VIACEP WEB SERVICE;
-Route::get("/viacep", [ViaCepController::class, "index"])->name("viacep.index");
-Route::post("/viacep", [ViaCepController::class, "index"])->name("viacep.index.post");
-Route::get("/viacep/{cep}", [ViaCepController::class, "show"])->name("viacep.show");
+Route::middleware(['auth'])->group(function (){
+    Route::get("/", [UserController::class, "index"])->name("users.index")->middleware('auth');
+    Route::get("/users/create", [UserController::class, "create"])->name("users.create")->middleware('auth');
+    Route::post("/users", [UserController::class, "store"])->name("users.store")->middleware('auth');
+    Route::put("/users/{id}", [UserController::class, "update"])->name("users.update")->middleware('auth');
+    Route::get("/users/{id}/edit", [UserController::class, "edit"])->name("users.edit")->middleware('auth');
+    Route::get("/users/{id}", [UserController::class, "show"])->name("users.show")->middleware('auth');
+    Route::delete("/users/{id}/delete", [UserController::class, "destroy"])->name("users.destroy")->middleware('auth');
+    
+    Route::get("/user/team/{id}", [TeamController::class, 'show'])->name('team.show')->middleware('auth');
+
+    Route::get('/posts', [PostController::class, 'index'])->name('posts.index')->middleware('auth');
+    Route::get('/users/{id}/posts', [PostController::class, 'show'])->name('posts.show')->middleware('auth');
+
+
+    Route::get("/viacep", [ViaCepController::class, "index"])->name("viacep.index")->middleware('auth');
+    Route::post("/viacep", [ViaCepController::class, "index"])->name("viacep.index.post")->middleware('auth');
+    Route::get("/viacep/{cep}", [ViaCepController::class, "show"])->name("viacep.show")->middleware('auth');
+});
